src: Update Gateway, move SettingsFolder#existenceStatus to Settings
